Burst what remain of utils into files
